refactor: load quantity discount widget via separate file
